added shopping list, updated ui, and changed how meals are scaled

+++ db/migrate/20120108050703_create_ingredients.rb
@@ -4,7 +4,7 @@ class CreateIngredients < ActiveRecord::Migration
       t.string :name
       t.text :description
       t.string :measurement_unit
-      t.integer :calories_per_unit
+      t.float :calories_per_unit
       t.string :food_group
 
       t.timestamps
